tests/requirements.txt
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
python-dotenv==1.0.1
google-cloud-firestore==2.15.0
